Une base saine améliorable qui résoud la plupart des instances à 20 avions testées. Facilement modifiables : méthode de sélection du prochain avion à affecter / heuristiques / priorité des noeuds

+++ src/a_star.hpp
@@ -9,7 +9,7 @@ typedef struct {
 	uint_fast8_t prof;
 	uint_fast8_t *etat;
 	int_fast8_t *indP;
-	int selection;
+	double selection;
 } noeud;
 
 class A_star {
